+++ AddOns/Platynator/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Platynator
 
-## [351](https://github.com/TheMouseNest/Platynator/tree/351) (2026-03-23)
-[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/350...351) [Previous Releases](https://github.com/TheMouseNest/Platynator/releases)
+## [352](https://github.com/TheMouseNest/Platynator/tree/352) (2026-03-24)
+[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/351...352) [Previous Releases](https://github.com/TheMouseNest/Platynator/releases)
 
-- [Fixes #272] Fix elite marker disappearing  
+- Fix "From you" toggle for "Crowd Control" auras  
+- Add options to hide health % symbol and customise combinations of health values